problem_7 - refactor

+++ src/main/scala/com/grobocop/euler/problem7/NthPrime.scala
@@ -1,12 +1,12 @@
 package com.grobocop.euler.problem7
 
 object NthPrime {
-  def apply(n: Long): Long = {
+  def apply(n: Int): Long = {
     lazy val primes: LazyList[Int] = 2 #:: LazyList.from(3)
       .filter(p => primes
         .takeWhile { x => x * x <= p }
         .forall(p % _ > 0))
-    primes.take(n.toInt)
+    primes.take(n)
       .last
   }
 }
